feat: 디렉토리 구조 개선 및 calendar, question, sideeffect 엔티티 구현
