+++ src/main/java/hanium/dongguk/question/domain/Question.java
@@ -1,0 +1,30 @@
+package hanium.dongguk.question.domain;
+
+import hanium.dongguk.global.base.BaseTimeEntity;
+import jakarta.persistence.*;
+import lombok.*;
+import org.hibernate.annotations.GenericGenerator;
+
+import java.util.UUID;
+
+@Entity
+@Table(name = "question")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Builder
+public class Question extends BaseTimeEntity {
+
+    @Id
+    @GeneratedValue(generator = "uuid2")
+    @GenericGenerator(name="uuid2", strategy = "uuid2")
+    @Column(name = "id", columnDefinition = "BINARY(16)", updatable = false, nullable = false)
+    private UUID id;
+
+    @Enumerated(EnumType.STRING)
+    @Column(name = "type", nullable = false)
+    private QuestionType type;   // 질문 타입 (Enum)
+
+    @Column(name = "answer", columnDefinition = "TEXT")
+    private String answer;  // 사용자가 작성한 답변
+}
