--- src/main/java/hanium/dongguk/user/mypage/domain/Calendar.java
@@ -1,37 +1,0 @@
-package hanium.dongguk.user.mypage.domain;
-
-import hanium.dongguk.global.base.BaseTimeEntity;
-import hanium.dongguk.user.patient.UserPatient;
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDate;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Builder
-public class Calendar extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY) // AUTO_INCREMENT
-    private Long id;
-
-    @Column(nullable = false)
-    private LocalDate date;
-
-    @Column(length = 500)
-    private String description;
-
-    @Enumerated(EnumType.STRING)
-    private Emotion emotion;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "question_id")
-    private Question question;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "patient_id")
-    private UserPatient patient;
-}
