--- src/main/java/hanium/dongguk/user/mypage/domain/SideEffect.java
@@ -1,24 +1,0 @@
-package hanium.dongguk.user.mypage.domain;
-
-import hanium.dongguk.global.base.BaseTimeEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Builder
-public class SideEffect extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY) // AUTO_INCREMENT
-    private Long id;
-
-    @Column(columnDefinition = "TEXT")
-    private String description;   // 부작용 내용
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "calendar_id", nullable = false)
-    private Calendar calendar;    // 어떤 기록(Calendar)에 속하는지
-}
